refactor: Moved extract_coordinates to utils
The function extract_coordinates() might be used in more places.
